requirements.txt
python-dotenv
yfinance
fredapi
pandas
